% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arse.R
\name{arse}
\alias{arse}
\title{Calculates the area of resilience to stress event below a baseline.}
\usage{
arse(data, xcoord, ycoord, ybase = NA, yinvert = FALSE,
  saveout = FALSE)
}
\arguments{
\item{data}{A dataframe with x- and y-coordinates in wide format.}

\item{xcoord}{A specified selection of the x-coordinate values within
the dataframe. The first x-coordinate value should correspond to the baseline
input of 'y'.}

\item{ycoord}{A specified selection of the y-coordinate values within the
dataframe. The first y-coordinate value should correspond to the baseline value
of 'y'. The second y-coordinate value should be the first measure of 'y' after
the intrusion of a stress event. The last value of 'y' should correspond to
the last measurement of 'y' over the measured timeframe.}

\item{ybase}{A specified selection of the baseline of the 'y' measured variable.
Users are advised to place baseline as the first instance of the y-coordinate
values. Function defaults to use the first y-coordinate value in the series.}

\item{yinvert}{Specifies whether resilience occurs above or below the baseline
depending on the meaning of high and low 'y' values. When parameter 'yinvert'
is set to 'FALSE' (the default), it is assumed that higher numbers are indicative
of positive (i.e., desired) 'y' values (e.g., exam grade). When 'yinvert' is set
to 'TRUE', it is assumed that lower numbers are indicative of positive
(i.e., desired) 'y' values (e.g., blood pressure).}

\item{saveout}{When the parameter 'saveout' is set to 'FALSE' (the default),
a vector of calculated arse values are given for each case. When 'saveout' is
set to 'TRUE', a dataframe of the original inputted dataset is returned with a
new column of calculated arse values.}
}
\value{
When the parameter 'saveout' is set to 'FALSE', a vector of calculated
arse values are given for each case. When 'saveout' is set to 'TRUE', a
dataframe of the original inputted dataset is returned with a new column of
calculated arse values.
}
\description{
This function takes a series of 'x,y' coordinates and a specified 'y'
baseline value. For a given set of x-coordinates over time and repeated
measures of a 'y' variable, this function calculates the area of resilience
to a stress event (arse) that is formed below the specified baseline value of
'y' using an implementation of the shoelace formula (Gauss's area formula)
for the area of irregular polygons. Any area above the baseline (i.e., growth)
is ignored. Smaller arse values are indicative of better resilience on the
measured variable.
}
\examples{
xc <- t(c(1,2,3,4,5,6,7,8,9,10))
yc <- t(c(75,53,37,25,27,35,50,75,75,75))
dataset1 <- data.frame(xc, yc)
arse(data = dataset1, xcoord = dataset1[,1:10], ycoord = dataset1[,11:20], saveout = TRUE)
}
