% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arsefigures.R
\docType{data}
\name{arsefigures}
\alias{arsefigures}
\title{arsefigures}
\format{A dataframe with 11 rows and 22 columns. There are 10 total examples.}
\usage{
bar
}
\description{
Dataset used to demonstrate different manifestations of arse. Cases correspond
to different figures from a forthcoming arse method paper. Figure number and panel
are noted #' in the last column. Additional data is also added to test other
situations that might occur with data sources (e.g., different baselines,
interpolation of points).
}
\keyword{datasets}
